Migrate Escalator tests for 8 (#8811)
* Migrate Escalator tests for 8

Tests require reindeer / base for escalator to render correctly.
