+++ uitest/src/main/java/com/vaadin/tests/components/grid/basics/EscalatorBasicClientFeatures.java
@@ -13,21 +13,26 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.v7.tests.components.grid.basicfeatures;
 
+package com.vaadin.tests.components.grid.basics;
+
+import com.vaadin.annotations.Theme;
+import com.vaadin.annotations.Title;
 import com.vaadin.annotations.Widgetset;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.tests.widgetset.TestingWidgetSet;
-import com.vaadin.tests.widgetset.client.v7.grid.EscalatorBasicClientFeaturesWidget;
+import com.vaadin.tests.widgetset.client.grid.EscalatorBasicClientFeaturesWidget;
 import com.vaadin.tests.widgetset.server.TestWidgetComponent;
 import com.vaadin.ui.UI;
 
 @Widgetset(TestingWidgetSet.NAME)
-public class EscalatorUpdaterUi extends UI {
+@Title("Escalator basic client features")
+@Theme("reindeer")
+public class EscalatorBasicClientFeatures extends UI {
 
     @Override
-    protected void init(VaadinRequest request) {
+    public void init(VaadinRequest request) {
         setContent(new TestWidgetComponent(
-                EscalatorBasicClientFeaturesWidget.UpdaterLifetimeWidget.class));
+                EscalatorBasicClientFeaturesWidget.class));
     }
 }
