+++ uitest/src/main/java/com/vaadin/tests/components/grid/basics/EscalatorUpdaterUi.java
@@ -13,24 +13,23 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
+package com.vaadin.tests.components.grid.basics;
 
-package com.vaadin.v7.tests.components.grid.basicfeatures;
-
-import com.vaadin.annotations.Title;
+import com.vaadin.annotations.Theme;
 import com.vaadin.annotations.Widgetset;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.tests.widgetset.TestingWidgetSet;
-import com.vaadin.tests.widgetset.client.v7.grid.EscalatorBasicClientFeaturesWidget;
+import com.vaadin.tests.widgetset.client.grid.EscalatorBasicClientFeaturesWidget;
 import com.vaadin.tests.widgetset.server.TestWidgetComponent;
 import com.vaadin.ui.UI;
 
 @Widgetset(TestingWidgetSet.NAME)
-@Title("Escalator basic client features")
-public class EscalatorBasicClientFeatures extends UI {
+@Theme("reindeer")
+public class EscalatorUpdaterUi extends UI {
 
     @Override
-    public void init(VaadinRequest request) {
+    protected void init(VaadinRequest request) {
         setContent(new TestWidgetComponent(
-                EscalatorBasicClientFeaturesWidget.class));
+                EscalatorBasicClientFeaturesWidget.UpdaterLifetimeWidget.class));
     }
 }
